### 2024-03-13 #6479 @editorialbot
[PRE REVIEW]: ncompare: A Python package for comparing netCDF structures
**Submitting author:** @dkauf42 (<a href="http://orcid.org/0000-0002-1487-7298">Daniel E Kaufman</a>)
**Repository:** https://github.com/nasa/ncompare
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.8.0
