### 2024-03-13 #6479 @arfon
@dkauf42 – thanks for your submission to JOSS. I see there is a NOSA license under the `licenses` folder. Could you also please make sure there is a plain-text version of the license at the root of the repository named `LICENSE`?
