### 2024-03-13 #6479 @arfon
Cross-linking with pyOpenSci submission here: https://github.com/pyOpenSci/software-submission/issues/146
