### 2024-03-15 #6479 @arfon
@RMeli – there's actually no work to do here as the paper has already been accepted at pyOpenSci. But thanks anyway!
