### 2024-03-15 #6479 @arfon
(@cmtso, @cmarmo – there's no work for you to do here as you did the review over in https://github.com/pyOpenSci/software-submission/issues/146 😄 )
