### 2024-04-10 #6490 @danielfromearth
My apologies for the delay, @arfon. The software has been approved to use the Apache License 2.0, so I will now be changing the license file completely.  I think this will be even better for the openness of this project. Will let you know when I have completed the changover.
