### 2024-04-11 nasa/ncompare#184 @danielfromearth
Change license
Will be changing to "Apache License 2.0".

Will help address issue brought up in the associated manuscript review of the Journal of Open Source Software (JOSS): <https://github.com/openjournals/joss-reviews/issues/6490>.
