### 2024-04-23 #6490 @danielfromearth
@arfon, although the license has been approved, there is still ongoing discussion regarding the copyright headers for source files. Sorry for the continued delay.
