### 2024-05-30 #6490 @danielfromearth
@arfon, thank you very much for your patience as the license and copyright questions were addressed!

`ncompare` — now bumped to version 1.9.0 — has been re-licensed under the Apache License 2.0, and a `LICENSE` file has been placed in the root directory of the repository. I hope that both addresses the license discoverability issue and improves the openness of this `ncompare` code.

Please let me know if there is anything else I can do to help move this submission review to the next steps.
