### 2024-06-02 #6490 @arfon
Thanks @danielfromearth – this sounds great. Can you make a new Zenodo release for this updated version please and let me know the DOI of that release here?
