### 2024-06-03 #6490 @danielfromearth
@arfon, I just created a new [release for version 1.9.0](https://github.com/nasa/ncompare/releases/tag/1.9.0) with a DOI of <[10.5281/zenodo.11448464](https://doi.org/10.5281/zenodo.11448464)>.
