### 2024-06-06 #6490 @danielfromearth
> 👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06490/joss.06490/10.21105.joss.06490.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#5460](https://github.com/openjournals/joss-papers/pull/5460), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

@arfon, the proof link doesn't seem to work here.  Does it work for you?
