### 2024-06-06 nasa/ncompare#229 @danielfromearth
Add JOSS info after acceptance
- Update `CITATION.cff` (as described in https://github.com/openjournals/joss-reviews/issues/6490#issuecomment-2152701795)
- Add JOSS badge to the `README.md` (as described in https://github.com/openjournals/joss-reviews/issues/6490#issuecomment-2152721911)
