### 2024-06-06 #6490 @danielfromearth
> @cmtso, @cmarmo – many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
>
> @danielfromearth – your paper is now accepted and published in JOSS ⚡🚀💥

Yay!  We really appreciate you dedicating time and effort toward reviewing this paper, and for your patience during the process!
